+++ src/query.js
@@ -180,7 +180,7 @@ const Logger = require('./logger');
 const CONVERSATION = 'Conversation';
 const MESSAGE = 'Message';
 const findConvIdRegex = new RegExp(
-  /^conversation.id\s*=\s*['"]((temp_)?layer:\/\/\/conversations\/.{8}-.{4}-.{4}-.{4}-.{12})['"]$/);
+  /^conversation.id\s*=\s*['"](layer:\/\/\/conversations\/.{8}-.{4}-.{4}-.{4}-.{12})['"]$/);
 
 class Query extends Root {
 
@@ -400,7 +400,7 @@ class Query extends Root {
 
       // We will already have a this._predicate if we are paging; else we need to extract the UUID from
       // the conversationId.
-      const uuid = (this._predicate || conversationId).replace(/^(temp_)?layer\:\/\/\/conversations\//, '');
+      const uuid = (this._predicate || conversationId).replace(/^layer\:\/\/\/conversations\//, '');
       if (uuid) {
         return {
           uuid,
@@ -446,8 +446,10 @@ class Query extends Root {
       // if (conversation && conversation.lastMessage && lastMessage && lastMessage.id === conversation.lastMessage.id) fromId = '';
       const newRequest = `conversations/${predicateIds.uuid}/messages?page_size=${pageSize}${fromId}`;
 
-      // Don't query on temporary ids, nor repeat still firing queries
-      if (!this._predicate.match(/temp_/) && newRequest !== this._firingRequest) {
+      // Don't query on unsaved Conversations, nor repeat still firing queries
+      // If there is no conversation, its presumed to still be on the server and waiting for the
+      // app to call `load()`.
+      if ((!conversation || conversation.isSaved()) && newRequest !== this._firingRequest) {
         this.isFiring = true;
         this._firingRequest = newRequest;
         this.client.xhr({
@@ -694,7 +696,7 @@ class Query extends Root {
   _handleConversationChangeEvent(evt) {
     let index = this._getIndex(evt.target.id);
 
-    // If its an ID change (from temp to non-temp id) make sure to update our data.
+    // If its an ID change (matching Distinct Conversation returned by server) make sure to update our data.
     // If dataType is an instance, its been updated for us.
     if (this.dataType === Query.ObjectDataType) {
       const idChanges = evt.getChangesFor('id');
@@ -868,6 +870,15 @@ class Query extends Root {
     }
   }
 
+  /**
+   * A Conversation ID changes if a matching Distinct Conversation was found on the server.
+   *
+   * If this Query's Conversation's ID has changed, update the predicate.
+   *
+   * @method _handleMessageConvIdChangeEvent
+   * @param {layer.LayerEvent} evt - A Message Change Event
+   * @private
+   */
   _handleMessageConvIdChangeEvent(evt) {
     const cidChanges = evt.getChangesFor('id');
     if (cidChanges.length) {
@@ -922,14 +933,18 @@ class Query extends Root {
 
   _handleMessageChangeEvent(evt) {
     let index = this._getIndex(evt.target.id);
-    const midChanges = evt.getChangesFor('id');
+    const positionChanges = evt.getChangesFor('position');
 
-    if (midChanges.length) {
-      if (this.dataType === Query.ObjectDataType) index = this._getIndex(midChanges[0].oldValue);
-      if (this._handleMessagePositionChange(evt, index)) return;
+    // If there are position changes, handle them.  If all the changes are position changes,
+    // exit when done.
+    if (positionChanges.length) {
+      if (this._handleMessagePositionChange(evt, index)) {
+        if (positionChanges.length === evt.changes.length) return;
+        index = this._getIndex(evt.target.id); // Get the updated position
+      }
     }
 
-    if (evt.target.conversationId === this._predicate && index !== -1) {
+    if (index !== -1) {
       if (this.dataType === Query.ObjectDataType) {
         this.data = [
           ...this.data.slice(0, index),